redesign scene logic
