+++ include/fountain/ft_time.h
@@ -33,6 +33,11 @@ public:
 	double secondsFromPause();
 	double secondsFromContinue();
 };
+
+}
+
+namespace fountain {
+extern ftTime::Clock mainClock;
 }
 
 #endif
